requirements.txt
requests
apscheduler
python-dotenv
yfinance
transformers
diffusers
torch
accelerate
pillow
